backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF>=1.25.0
python-docx==1.1.2
pdf2image==1.17.0
reportlab==4.2.2
pydantic==2.9.0
python-dotenv>=1.0.0
pyjwt[crypto]>=2.8.0
httpx>=0.25.0
